+++ Sources/Modules/Rendering/Resources/GPUSwapchain.hpp
@@ -11,7 +11,6 @@ struct GPUSwapchainData final : ResourceData
 {
 	vk::raii::SwapchainKHR			 swapchain;
 	std::vector<vk::raii::ImageView> images_views;
-	std::vector<vk::raii::Semaphore> present_semaphores;
 
 	#pragma region Lifetime
